Fix the following errors due to some changes in the implementation of some components. Rename the LabelViewTest to RoleViewTest. Also, fix errors due to changes in its input properties. Move ShowTeamModalTest to the folder of its component. Also, fix errors due to some name changes on the input properties of ShowTeamModal. Move TaskInputTest to the folder of its component. Fix errors due to changes on the naming of the TaskView’s input properties.
